+++ clients/include/hipsolver_dispatcher.hpp
@@ -13,6 +13,7 @@
 #include "testing_getrf.hpp"
 #include "testing_getrf_npvt.hpp"
 #include "testing_getrs.hpp"
+#include "testing_orgqr_ungqr.hpp"
 #include "testing_ormqr_unmqr.hpp"
 #include "testing_potrf.hpp"
 #include "testing_sytrd_hetrd.hpp"
@@ -60,6 +61,7 @@ class hipsolver_dispatcher
     {
         // Map for functions that support single and double precisions
         static const func_map map_real = {
+            {"orgqr", testing_orgqr_ungqr<false, T>},
             {"ormqr", testing_ormqr_unmqr<false, T>},
             {"sytrd", testing_sytrd_hetrd<false, false, false, T>},
         };
@@ -80,6 +82,7 @@ class hipsolver_dispatcher
     {
         // Map for functions that support single complex and double complex precisions
         static const func_map map_complex = {
+            {"ungqr", testing_orgqr_ungqr<false, T>},
             {"unmqr", testing_ormqr_unmqr<false, T>},
             {"hetrd", testing_sytrd_hetrd<false, false, false, T>},
         };