+++ clients/include/hipsolver_fortran.hpp
@@ -22,6 +22,75 @@ extern "C" {
  *   LAPACK
  * ========== */
 
+// orgqr/ungqr
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverSorgqr_bufferSizeFortran(
+    hipsolverHandle_t handle, int m, int n, int k, float* A, int lda, float* tau, int* lwork);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDorgqr_bufferSizeFortran(
+    hipsolverHandle_t handle, int m, int n, int k, double* A, int lda, double* tau, int* lwork);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverCungqr_bufferSizeFortran(hipsolverHandle_t handle,
+                                                                     int               m,
+                                                                     int               n,
+                                                                     int               k,
+                                                                     hipsolverComplex* A,
+                                                                     int               lda,
+                                                                     hipsolverComplex* tau,
+                                                                     int*              lwork);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverZungqr_bufferSizeFortran(hipsolverHandle_t       handle,
+                                                                     int                     m,
+                                                                     int                     n,
+                                                                     int                     k,
+                                                                     hipsolverDoubleComplex* A,
+                                                                     int                     lda,
+                                                                     hipsolverDoubleComplex* tau,
+                                                                     int*                    lwork);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverSorgqrFortran(hipsolverHandle_t handle,
+                                                          int               m,
+                                                          int               n,
+                                                          int               k,
+                                                          float*            A,
+                                                          int               lda,
+                                                          float*            tau,
+                                                          float*            work,
+                                                          int               lwork,
+                                                          int*              devInfo);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDorgqrFortran(hipsolverHandle_t handle,
+                                                          int               m,
+                                                          int               n,
+                                                          int               k,
+                                                          double*           A,
+                                                          int               lda,
+                                                          double*           tau,
+                                                          double*           work,
+                                                          int               lwork,
+                                                          int*              devInfo);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverCungqrFortran(hipsolverHandle_t handle,
+                                                          int               m,
+                                                          int               n,
+                                                          int               k,
+                                                          hipsolverComplex* A,
+                                                          int               lda,
+                                                          hipsolverComplex* tau,
+                                                          hipsolverComplex* work,
+                                                          int               lwork,
+                                                          int*              devInfo);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverZungqrFortran(hipsolverHandle_t       handle,
+                                                          int                     m,
+                                                          int                     n,
+                                                          int                     k,
+                                                          hipsolverDoubleComplex* A,
+                                                          int                     lda,
+                                                          hipsolverDoubleComplex* tau,
+                                                          hipsolverDoubleComplex* work,
+                                                          int                     lwork,
+                                                          int*                    devInfo);
+
 // ormqr/unmqr
 HIPSOLVER_EXPORT hipsolverStatus_t hipsolverSormqr_bufferSizeFortran(hipsolverHandle_t    handle,
                                                                      hipsolverSideMode_t  side,